+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/shippinglabels/creation/EditShippingLabelPackagesViewModel.kt
@@ -87,7 +87,7 @@ class EditShippingLabelPackagesViewModel @AssistedInject constructor(
         return listOf(
             ShippingLabelPackage(
                 selectedPackage = lastUsedPackage,
-                weight = Double.NaN,
+                weight = Float.NaN,
                 items = order.getShippableItems().map { it.toShippingItem() }
             )
         )
@@ -108,7 +108,7 @@ class EditShippingLabelPackagesViewModel @AssistedInject constructor(
         }
     }
 
-    fun onWeightEdited(position: Int, weight: Double) {
+    fun onWeightEdited(position: Int, weight: Float) {
         val packages = viewState.shippingLabelPackages.toMutableList()
         packages[position] = packages[position].copy(weight = weight)
         viewState = viewState.copy(shippingLabelPackages = packages)
